[ADD] get projects by user

+++ src/api/components/projects/controller.js
@@ -25,14 +25,41 @@ const saveProject = (title, description, technologies, rol, repository, url, use
 
 const deleteOne = async (id, userData) => {
     const project = await store.getOneById({_id: id});
-    console.log(userData.sub, project.user['_id']);
+    if(!project) throw boom.badData('Cannot find a project with that id');
     if(userData.sub != project.user['_id']) throw boom.unauthorized('You are not authorized to do that action.');
-
     return store.deleteOne(id);
+}
+
+const updateProject = async (id, title, description, technologies, rol, repository, url, userData) => {
+    const project = await store.getOneById({_id: id});
+    if(!project) throw boom.badData('Cannot find a project with that id');
+    if(userData.sub != project.user['_id']) throw boom.unauthorized('You are not authorized to do that action.');
+    let update = {
+        id,
+        title,
+        description,
+        technologies,
+        rol,
+        repository,
+        url,
+    }
+    for(let prop in update) {
+        if(!update[prop]) delete update[prop];
+    }
+    return await store.update(id, update);
+}
+
+const getProjectsOf = async (id) => {
+    let filter = {
+        user: id
+    }
+    return await store.getProjectsByFilter(filter);
 }
 
 module.exports = {
     saveProject,
     getOneById,
-    deleteOne
+    deleteOne,
+    updateProject,
+    getProjectsOf
 }
